add Hmily SQLManipulation enum
