+++ hmily-repository/hmily-repository-spi/src/main/java/org/dromara/hmily/repository/spi/entity/HmilyDataSnapshot.java
@@ -19,6 +19,7 @@ package org.dromara.hmily.repository.spi.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.dromara.hmily.repository.spi.entity.tuple.HmilySQLTuple;
 
 import java.io.Serializable;
 import java.util.Collection;
